requirements.txt
opencv-python
numpy


rembg
Pillow
onnxruntime
